requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
discord==2.3.2
discord.py==2.4.0
frozenlist==1.4.1
idna==3.7
multidict==6.0.5
numpy==2.0.1
opencv-python==4.10.0.84
packaging==24.1
pillow==10.4.0
pytesseract==0.3.10
yarl==1.9.4
